requirements.txt
appdirs==1.4.4
argcomplete==3.1.4
asgiref==3.6.0
Brotli==1.0.9
cffi==1.15.1
colorama==0.4.6
cssselect2==0.7.0
distlib==0.3.6
Django==4.2.7
django-crum==0.7.9
escpos==2.0.0
et-xmlfile==1.1.0
filelock==3.10.0
fonttools==4.39.4
future==0.18.3
gunicorn==21.2.0
html5lib==1.1
image==1.5.33
itsdangerous==2.1.2
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
Pillow==9.4.0
platformdirs==3.1.1
psycopg2==2.9.5
pycparser==2.21
pydyf==0.6.0
pyphen==0.14.0
pypng==0.20220715.0
pyserial==3.5
python-barcode==0.15.1
python-dotenv==1.0.0
pyusb==1.2.1
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.12
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2022.7
viivakoodi==0.8.0
virtualenv==20.21.0
waitress==2.1.2
weasyprint==58.1
webencodings==0.5.1
wfastcgi==3.0.0
whitenoise==6.6.0
zopfli==0.2.2
